+++ api/src/modules/notification/notification.module.ts
@@ -1,4 +1,4 @@
-import { Module, Provider } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { UserModule } from 'src/modules/user/user.module';
@@ -7,16 +7,10 @@ import { NotificationController } from './notification.controller';
 import { Notification } from './notification.entity';
 import { NotificationService } from './notification.service';
 
-const NOTIFICATION_PAGE_SIZE = 'NOTIFICATION_PAGE_SIZE';
-const NotificationPageSize: Provider = {
-  provide: NOTIFICATION_PAGE_SIZE,
-  useValue: 10,
-};
-
 @Module({
   imports: [TypeOrmModule.forFeature([Notification]), UserModule],
   controllers: [NotificationController],
-  providers: [NotificationPageSize, NotificationService],
+  providers: [NotificationService],
   exports: [NotificationService],
 })
 export class NotificationModule {}